Modify Jenkinsfile.groovy

+++ Jenkinsfile.groovy
@@ -28,10 +28,12 @@ pipeline {
         stage('Connect Deploy Server') {
             steps {
                 sh 'sshpass -p $DEPLOY_SERVER_PASSWORD ssh -T -p 12308 $DEPLOY_SERVER_ID@106.10.59.248'
-                PREVIOUS_PID=$(ps -ef | grep java | grep jujeol | grep -v nohup | awk 'build')
+                PREVIOUS_PID=sh "ps -ef | grep java | grep build/fire_inform-0.0.1-SNAPSHOT.jar | grep -v nohup | awk '{print $2}'"
                 echo '$PREVIOUS_PID'
-                sh 'pkill -9 -ef build/fire_inform-0.0.1-SNAPSHOT.jar'
-                sh 'sleep 5'
+                if(${PREVIOUS_PID} != null) {
+                    sh 'pkill -9 -ef build/fire_inform-0.0.1-SNAPSHOT.jar'
+                    sh 'sleep 5'
+                }
                 sh 'nohup java -jar build/fire_inform-0.0.1-SNAPSHOT.jar &'
             }
         }
